Create system/keycap compatibility layer
These layers map the (assumed) DE system language to the actual keycaps
(which are QWERTY ANSI).

While most of the time the actual keycap labels are not important,
because of touch typing, some special characters might need reference.
For this use case, these layer map the actual labels (of QWERTY ANSI
caps) to german system input keycodes (as the keyboard will be used
exclusively with german system settings; and even if the system is
running in en_US: The base layer will work with the caps).

This might be considered a pure convenience layer. Probably it might be
actively used during programming, as some brackets are easier to reach
on QWERTY ANSI.